Moved out all nested atoms
