--- app/components/UserProfile/atoms/ProfileInfo/atoms/style.css
@@ -1,52 +1,0 @@
-.ProfileAvatar {
-  width: 5.5vw;
-  position: relative;
-  float: right;
-  text-align: center;
-}
-.ProfileAvatar__image {
-  width: 100%;
-  height: 100%;
-}
-.ProfileDescription__name {
-  font-family: 'PlayfairDisplay';
-  font-style: normal;
-  font-weight: bold;
-  font-size: 2rem;
-  line-height: 40px;
-  margin-bottom: 0px;
-}
-.ProfileDescription__company {
-  font-family: 'Open-Sans';
-  font-style: normal;
-  font-weight: 300;
-  font-size: 1rem;
-  line-height: 19px;
-  margin-bottom: 0.5rem;
-  color: rgba(42, 42, 42, 0.7);
-}
-.ProfileContacts {
-  margin-bottom: 1rem;
-}
-.ProfileContacts__email,
-.ProfileContacts__phone {
-  color: rgba(42, 42, 42, 0.7);
-  margin-right: 1rem;
-  text-decoration: underline;
-}
-.ProfileInfo__about {
-  width: 65%;
-  font-family: 'Open-Sans';
-  font-style: normal;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 22px;
-
-  color: #2a2a2a;
-}
-
-.ProfileAvatar__camera {
-  position: absolute;
-  bottom: 0;
-  right: 0;
-}
